Aggregate api create football api repository (#144)
* Configured FootballAPIStanding in  appsettings and created FootballAPI_HttpClient in Infrastructure layer.

* Created The FootballStandingsService and registered in the DI container.

Also Created an HttpGet method in the the Controller that exposes the GetFootbalStandingsAsync method.

Renamed method in the IFootballAPI_HttpClient from GetFootballData to GetFootballDataAsync.

* Made corrections to the ResponseApi model class as it was causing a discrepancy between the incoming json response structure and the model meant to acomodate it. Also registered FootballStandings API headers with the HttpClient in the infrastructure DI container.

* Added Logging to the  FootballStandingsService and FootballAPI_HttpClient.

Used HttpClientFactory instead of HttpClient in FootballAPI_HttpClient in order to get the headers from the HttpClient registration in the Inframodule Di container. This acts as centralised location to configure named clients.

Ensured that both serialization and deserialization use the same JsonSerializerOptions in RedisCache to maintain consistency.